(feat) Limit board deletion to user's own boards

+++ server/orm/controllers/BoardController.js
@@ -38,15 +38,20 @@ export default {
 
   deleteBoard: (req, res) => {
     const id = req.params.board_id;
+    const userId = req.user.sub;
 
     Board.get(id).getJoin({
       ideas: true,
     }).run()
       .then((board) => {
-        board.deleteAll({ ideas: true })
-          .then((result) => {
-            res.sendStatus(204);
-          });
+        if (userId === board.authorId) {
+          board.deleteAll({ ideas: true })
+            .then((result) => {
+              res.sendStatus(204);
+            });
+        } else {
+          console.log('Permission denied.');
+        }
       })
       .error(helper.handleError(res));
   },
